pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.10.1",
    "crewai-tools>=1.10.1",
    "google-genai>=1.67.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.1.1",
]
